+++ example/main.dart
@@ -33,6 +33,7 @@ const String _kJobsOption = 'jobs';
 const String _kWorkingDirectoryOption = 'working-directory';
 const String _kCommandOption = 'command';
 const String _kSourceOption = 'source';
+const String _kGroupOption = 'group';
 const String _kAppName = 'process_runner';
 
 // This only works for escaped spaces and things in double or single quotes.
@@ -198,6 +199,17 @@ Future<void> main(List<String> args) async {
         'properly on the command line.',
   );
   parser.addMultiOption(
+    _kGroupOption,
+    abbr: 'g',
+    defaultsTo: <String>[],
+    help: 'Specify the name of a file to read grouped commands from, one per '
+        'line, as they would appear on the command line, with spaces escaped '
+        'or quoted. More than one --$_kGroupOption argument may be specified, '
+        'and each file will be run as a separate job group with the commands '
+        'in each group run in the order specified in the file, but in parallel '
+        'with other groups.',
+  );
+  parser.addMultiOption(
     _kSourceOption,
     abbr: 's',
     defaultsTo: <String>[],
@@ -205,8 +217,8 @@ Future<void> main(List<String> args) async {
         'they would appear on the command line, with spaces escaped or '
         'quoted. Specify "--$_kSourceOption -" to read from stdin. More than '
         'one --$_kSourceOption argument may be specified, and they will be '
-        'concatenated in the order specified. The stdin ("--$_kSourceOption -") '
-        'argument may only be specified once.',
+        'run as separate tasks in parallel. The stdin '
+        '("--$_kSourceOption -") argument may only be specified once.',
   );
 
   late ArgResults options;
@@ -225,6 +237,7 @@ Future<void> main(List<String> args) async {
       '[--$_kWorkingDirectoryOption=<working directory>] '
       '[--$_kJobsOption=<num_worker_jobs>] '
       '[--$_kCommandOption="command" ...] '
+      '[--$_kGroupOption=<file> ...] '
       '[--$_kSourceOption=<file|"-"> ...]:',
     );
 
@@ -242,47 +255,31 @@ Future<void> main(List<String> args) async {
 
   // Collect the commands to be run from the command file(s).
   final List<String>? commandFiles = options[_kSourceOption] as List<String>?;
-  final List<String> fileCommands = <String>[];
-  if (commandFiles != null) {
-    bool sawStdinAlready = false;
-    for (final String commandFile in commandFiles) {
-      // Read from stdin if the --file option is set to '-'.
-      if (commandFile == '-') {
-        if (sawStdinAlready) {
-          stderr.writeln('ERROR: The stdin can only be specified once with "--$_kSourceOption -"');
-          exitCode = 1;
-          return;
-        }
-        sawStdinAlready = true;
-        String? line = stdin.readLineSync();
-        while (line != null) {
-          fileCommands.add(line);
-          line = stdin.readLineSync();
-        }
-      } else {
-        // Read the commands from a file.
-        final File cmdFile = File(commandFile);
-        if (!cmdFile.existsSync()) {
-          print('''Command file "$commandFile" doesn't exist.''');
-          exit(1);
-        }
-        fileCommands.addAll(cmdFile.readAsLinesSync());
-      }
-    }
-  }
-
-  // Collect all the commands, both from the input file, and from the command
-  // line. The command line commands come first (although they could all
+  // Collect all the commands, both from input files, and from the command
+  // line. The command line commands are run first (although they could all
   // potentially be executed simultaneously, depending on the number of workers,
   // and number of commands).
-  final List<String> commands = <String>[
+  final List<List<String>> fileCommands = getCommandsFromFiles(commandFiles);
+
+  final List<String> collectedCommands = <String>[
     if (options[_kCommandOption] != null) ...options[_kCommandOption]! as List<String>,
-    ...fileCommands,
   ];
+  fileCommands
+      .forEach(collectedCommands.addAll); // Flatten the groups so they can be run in parallel.
+
+  final List<List<String>> splitCommands =
+      collectedCommands.map<List<String>>((String command) => splitIntoArgs(command)).toList();
+
+  // Collect the commands to be run from the group file(s).
+  final List<List<String>> groupCommands =
+      getCommandsFromFiles(options[_kGroupOption] as List<String>, allowStdin: false);
 
   // Split each command entry into a list of strings, taking into account some
   // simple quoting and escaping.
-  final List<List<String>> splitCommands = commands.map<List<String>>(splitIntoArgs).toList();
+  final List<List<List<String>>> splitGroupCommands = groupCommands
+      .map<List<List<String>>>(
+          (List<String> group) => group.map<List<String>>(splitIntoArgs).toList())
+      .toList();
 
   // If the numWorkers is set to null, then the ProcessPool will automatically
   // select the number of processes based on how many CPU cores the machine has.
@@ -294,14 +291,25 @@ Future<void> main(List<String> args) async {
     numWorkers: numWorkers,
     printReport: printReport ? stderrPrintReport : null,
   );
-  final List<WorkerJob> jobs = splitCommands.map<WorkerJob>((List<String> command) {
-    return WorkerJob(
-      command,
-      workingDirectory: workingDirectory,
-      runInShell: runInShell,
-      failOk: failOk,
-    );
-  }).toList();
+  final Iterable<WorkerJobGroup> groupedJobs =
+      splitGroupCommands.map<WorkerJobGroup>((List<List<String>> group) {
+    return WorkerJobGroup(group
+        .map<WorkerJob>((List<String> command) => WorkerJob(
+              command,
+              workingDirectory: workingDirectory,
+              runInShell: runInShell,
+              failOk: failOk,
+            ))
+        .toList());
+  });
+  final Iterable<WorkerJob> parallelJobs =
+      splitCommands.map<WorkerJob>((List<String> command) => WorkerJob(
+            command,
+            workingDirectory: workingDirectory,
+            runInShell: runInShell,
+            failOk: failOk,
+          ));
+  final Iterable<DependentJob> jobs = <DependentJob>[...parallelJobs, ...groupedJobs];
   try {
     await for (final WorkerJob done in pool.startWorkers(jobs)) {
       if (printStdout) {
@@ -321,4 +329,38 @@ Future<void> main(List<String> args) async {
 
   // Return non-zero exit code if there were jobs that failed.
   exitCode = pool.failedJobs != 0 ? 1 : 0;
+}
+
+List<List<String>> getCommandsFromFiles(List<String>? commandFiles, {bool allowStdin = false}) {
+  final List<List<String>> fileCommands = <List<String>>[];
+  if (commandFiles != null) {
+    bool sawStdinAlready = false;
+    for (final String commandFile in commandFiles) {
+      // Read from stdin if the --file option is set to '-'.
+      if (allowStdin && commandFile == '-') {
+        if (sawStdinAlready) {
+          stderr.writeln('ERROR: The stdin can only be specified once with "--$_kSourceOption -"');
+          exitCode = 1;
+          return <List<String>>[];
+        }
+        sawStdinAlready = true;
+        String? line = stdin.readLineSync();
+        final List<String> commands = <String>[];
+        while (line != null) {
+          commands.add(line);
+          line = stdin.readLineSync();
+        }
+        fileCommands.add(commands);
+      } else {
+        // Read the commands from a file.
+        final File cmdFile = File(commandFile);
+        if (!cmdFile.existsSync()) {
+          print('''Command file "$commandFile" doesn't exist.''');
+          exit(1);
+        }
+        fileCommands.add(cmdFile.readAsLinesSync());
+      }
+    }
+  }
+  return fileCommands;
 }